Updates RanchServer to use a Protocol Serializer on connect
In order to generalize the RanchServer, it has been updated to accept a
serializer via config options. If one is not provided, it defaults to using
Poison. This is needed in order to parse the path and select a handler which
will then use the serializer provided by the Transport for all other messages.